+++ CHANGELOG.md
@@ -5,7 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-# Fixed
+### Added
+
+- Added OpenID Connect `StateStore` component at `openidconnect/http/statestore`.
+
+### Fixed
 
 - `repository` and `bugs` URLs in `package.json` pointed to correct GitHub
 repository.